finder fixes and url_title enhancements
Finder module can now handle connection errors.
Url_title module now only gets the title if someone asked for it
Url_title has a timeout for the request, so the bot is looked for a
maximum of 5 seconds.
Url_title only outputs the title of the first url in a message to avoid
locking up the bot for a long period of time.
Url_title no longer uses the persistent storage.